perf(web): code-split /discover below-the-fold rows + 5 hooks (refs #979)
Follow-up to PR #1211 bundle-budget bump. Extracts Rows 2-7 (newGames,
popularAgents, recommendedToolkits, recentKbDocs, topContributors, events)
+ FooterCTA + their 5 useDiscover* hooks + adapters into a new
`_DiscoverBelowFoldRows.tsx` wrapper, lazy-loaded via next/dynamic with
ssr:false (TanStack Query is client-only).

Row 1 (trending) stays in page.tsx because it sits above the fold on all
viewports — it must paint synchronously with the Hero.

Pattern reference: apps/web/src/app/(authenticated)/play-records/page.tsx:18-21

Bundle impact (gzipped, measured locally with `pnpm bundle:check`):
- /discover: 636.7 KB → 635.6 KB  (saving ~1.1 KB, defensive split)
- Other 4 budgeted routes: unchanged (no shared imports touched)

Chunk verification: `_DiscoverBelowFoldRows` body lands in a separate chunk
(0dvxeutueook9.js) not part of /discover firstLoadChunkPaths — confirming
the dynamic import boundary is honored by Turbopack.

The primary value here is defensive — adding new sub-features inside the
below-the-fold block (e.g. more rows, new hooks, schemas) will accrue to
the lazy chunk instead of inflating the parent route bundle. Combined with
the .bundle-budgets.json 2026-05-17 baseline (#1211), this should keep
/discover within the 5% tolerance through the next 2-3 feature rounds.

Implementation notes:
- `EVENTS_ENDPOINT_AVAILABLE = false` constant moved into the wrapper since
  Row 7 (events disabled-shell) is the only consumer.
- `rowVisible()` callback duplicated as a local pure function in the wrapper
  (4 lines) to avoid passing a callback prop; trade-off chosen for chunk
  isolation.
- Inline `<BelowFoldSkeleton>` component (~25 LOC) in page.tsx reserves
  ~1400px of vertical space to prevent CLS during lazy chunk load. Uses
  semantic tokens (bg-muted/30, bg-muted/40) per Token Canonicalization rules.

No route behavior changes; the /discover render tree is identical, only
the JavaScript loading sequence is split. No new tests required (no e2e
or unit tests existed for /discover; the existing manual telemetry events
fire on the same triggers as before).
